ENG-11055: review changes
