+++ lib/python/voltcli/voltadmin.d/stop.py
@@ -17,63 +17,8 @@
 # Stop a node. Written to easily support multiple, but configured for
 # a single host for now.
 
-
-class Host(dict):
-
-    def __init__(self, id, abort_func):
-        self.id = id
-        self.abort_func = abort_func
-
-    # Provide pseudo-attributes for dictionary items with error checking
-    def __getattr__(self, name):
-        try:
-            return self[name]
-        except IndexError:
-            self.abort_func('Attribute "%s" not present for host.' % name)
-
-    def get_admininterface(self):
-        """
-        Return the likely admininterface.
-        Implementation note: Currently, ipaddress will be set to externalinterface
-        if set by user, otherwise server selects any interface.
-        """
-
-        if 'admininterface' in self and self['admininterface']:
-            return self['admininterface']
-        if 'ipaddress' in self and self['ipaddress']:
-            return self['ipaddress']
-        return None
-
-class Hosts(object):
-
-    def __init__(self, abort_func):
-        self.hosts_by_id = {}
-        self.abort_func = abort_func
-
-    def update(self, host_id_raw, prop_name_raw, value_raw):
-        host_id = int(host_id_raw)
-        prop_name = prop_name_raw.lower()
-        value = value_raw
-        if prop_name.endswith('port'):
-            value = int(value)
-        self.hosts_by_id.setdefault(host_id, Host(host_id, self.abort_func))[prop_name] = value
-
-    def get_target_and_connection_host(self, host_name):
-        """
-        Find an arbitrary host that isn't the one being stopped.
-        Returns a tuple with connection and target host objects.
-        """
-        connection_host = None
-        target_host = None
-        for host in self.hosts_by_id.values():
-            if host.hostname == host_name:
-                target_host = host
-            elif connection_host is None:
-                connection_host = host
-            if not connection_host is None and not target_host is None:
-                break
-        return (target_host, connection_host)
-
+from voltcli.hostinfo import Host
+from voltcli.hostinfo import Hosts
 
 @VOLT.Command(
     bundles = VOLT.AdminBundle(),
